Move cleanup_org_cache_dirs. Rename BaseProjectConfig.py
