+++ cumulusci/core/keychain/__init__.py
@@ -1,8 +1,10 @@
 # IMPORT ORDER MATTERS!
 
 # inherit from BaseConfig
-from cumulusci.core.keychain.BaseProjectKeychain import BaseProjectKeychain
-from cumulusci.core.keychain.BaseProjectKeychain import DEFAULT_CONNECTED_APP
+from cumulusci.core.keychain.base_project_keychain import (
+    BaseProjectKeychain,
+    DEFAULT_CONNECTED_APP,
+)
 
 # inherit from BaseProjectKeychain
 from cumulusci.core.keychain.BaseEncryptedProjectKeychain import (
