+++ cumulusci/core/keychain/base_project_keychain.py
@@ -8,7 +8,7 @@ from cumulusci.core.exceptions import OrgNotFound
 from cumulusci.core.exceptions import ServiceNotConfigured
 from cumulusci.core.exceptions import ServiceNotValid
 from cumulusci.core.sfdx import sfdx
-from cumulusci.utils import fileutils
+from cumulusci.core.utils import cleanup_org_cache_dirs
 
 DEFAULT_CONNECTED_APP = ConnectedAppOAuthConfig(
     {
@@ -137,7 +137,7 @@ class BaseProjectKeychain(BaseConfig):
     def remove_org(self, name, global_org=None):
         if name in self.orgs.keys():
             self._remove_org(name, global_org)
-        fileutils.cleanup_org_cache_dirs(self, self.project_config)
+        cleanup_org_cache_dirs(self, self.project_config)
 
     def _remove_org(self, name, global_org):
         del self.orgs[name]
