+++ cumulusci/core/utils.py
@@ -9,6 +9,7 @@ import copy
 import glob
 import pytz
 import time
+from shutil import rmtree
 
 from cumulusci.core.exceptions import ConfigMergeError, TaskOptionsError
 
@@ -178,3 +179,23 @@ def dictmerge(a, b, name=None):
             config_name=name,
         )
     return a
+
+
+def cleanup_org_cache_dirs(keychain, project_config):
+    """Cleanup directories that are not associated with a connected/live org."""
+    domains = set()
+    for org in keychain.list_orgs():
+        org_config = keychain.get_org(org)
+        domain = org_config.get_domain()
+        if domain:
+            domains.add(domain)
+
+    assert project_config.project_cache_dir
+    assert keychain.global_config_dir
+
+    project_org_directories = (project_config.project_cache_dir / "orginfo").glob("*")
+    global_org_directories = (keychain.global_config_dir / "orginfo").glob("*")
+
+    for directory in list(project_org_directories) + list(global_org_directories):
+        if directory.name not in domains:
+            rmtree(directory)
